requirements.txt
alembic==0.8.7
banal==1.0.1
bleach==1.4.3
dataset==1.4.4
feedparser==5.2.1
html5lib==0.9999999
Mako==1.1.4
MarkupSafe==0.23
normality==0.2.4
oauthlib==3.0.0
Pillow==7.1.1
PySocks==1.7.1
python-editor==1.0.1
pytz==2016.6.1
PyYAML==5.4.1
requests==2.11.1
requests-oauthlib==1.3.0
selenium==2.53.6
simplediff==1.0
six==1.10.0
SQLAlchemy==1.3.6
tweepy==3.10.0
beautifulsoup4==4.9.3
flake8==3.8.4